pyproject.toml
[project]
name = "accountant-bot"
version = "0.1.0"
description = "Telegram bot for accounting"
authors = [{ name = "Maximilian Hagelstam" }]
readme = "README.md"
requires-python = ">=3.14"
classifiers = [
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "python-telegram-bot>=22.1",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.2.1",
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "mypy>=0.991",
    "pytest-cov>=7.0.0",
    "ruff>=0.11.5",
]

[project.urls]
Homepage = "https://github.com/hagelstam/accountant-bot"
Repository = "https://github.com/hagelstam/accountant-bot"
Documentation = "https://github.com/hagelstam/accountant-bot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/accountant_bot"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
